Refactor import paths
