+++ pyproject.toml
@@ -1,71 +1,63 @@
-[build-system]
-requires      = ["setuptools>=61.0.0", "wheel"]
-build-backend = "setuptools.build_meta"
-
-[project]
+[tool.poetry]
 name = "worldenergydata"
 version = "0.0.1"
 description = "World Energy Data from Public Sources"
+authors = ["Vamsee Achanta <vamsee.achanta@aceengineer.com>"]
+license = "MIT"
 readme = "README.md"
-authors = [{ name = "Vamsee Achanta", email = "vamsee.achanta@aceengineer.com" }]
-license = { file = "LICENSE" }
-classifiers = [
-    "License :: OSI Approved :: MIT License",
-    "Programming Language :: Python",
-    "Programming Language :: Python :: 3",
-]
-keywords=['package', 'library', 'module']
-dependencies = ["pyyaml", 'tomli; python_version < "3.11"']
+keywords = ["package", "library", "module"]
+homepage = "https://github.com/vamseeachanta/worldenergydata"
+repository = "https://github.com/vamseeachanta/worldenergydata"
+documentation = "https://github.com/vamseeachanta/worldenergydata/blob/main/README.md"
+packages = [{ include = "worldenergydata", from = "src" }]
+include = ["*.yml", "*.csv", "*.xlsx", "*.html", "*.txt"]
 requires-python = ">=3.7"
 
-    [project.optional-dependencies]
-    build = ["build", "twine"]
-    dev   = ["black", "bumpver", "isort", "pytest"]
+[tool.poetry.dependencies]
+python = ">=3.7"
 
-    [project.scripts]
-    energy_data = "worldenergydata:__main__"
+[tool.poetry.group.build.dependencies]
+build = "^1.0.0"
+twine = "^1.0.0"
 
-    [project.urls]
-    repository    = "https://github.com/vamseeachanta/worldenergydata"
-    documentation = "https://github.com/vamseeachanta/worldenergydata/blob/main/README.md"
+[tool.poetry.group.dev.dependencies]
+black = "^23.0"
+bumpver = "^2023.1129"
+isort = "^5.0.0"
+pytest = "^7.0.0"
 
-[tool.setuptools]
-include-package-data = true
+[tool.poetry.scripts]
+energy_data = "worldenergydata.__main__:main"
 
-[tool.setuptools.dynamic]
-dependencies = {file = ["dev_tools/requirements.txt"]}
-
-[tool.setuptools.packages.find]
-where = ["src"]
-[tool.setuptools.package-data]
-"*" = ["*.yml", "*.csv", "*.xlsx", "*.html", "*.txt"]
-
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
 
 [tool.bumpver]
 current_version = "0.0.1"
 version_pattern = "MAJOR.MINOR.PATCH"
-commit_message  = "bump version {old_version} -> {new_version}"
-commit          = true
-tag             = true
-push            = false
+commit_message = "bump version {old_version} -> {new_version}"
+commit = true
+tag = true
+push = false
 
-    [tool.bumpver.file_patterns]
-    "pyproject.toml" = [
-        'current_version = "{version}"',
-        'version = "{version}"',
-    ]
-    "src/worldenergydata/__init__.py" = ["{version}"]
-    "src/worldenergydata/__main__.py" = ["- worldenergydata v{version}"]
+[tool.bumpver.file_patterns]
+"pyproject.toml" = [
+    'current_version = "{version}"',
+    'version = "{version}"',
+]
+"src/worldenergydata/__init__.py" = ["{version}"]
+"src/worldenergydata/__main__.py" = ["- worldenergydata v{version}"]
 
 [tool.isort]
-profile                   = "black"
-import_heading_stdlib     = "Standard library imports"
+profile = "black"
+import_heading_stdlib = "Standard library imports"
 import_heading_thirdparty = "Third party imports"
 import_heading_firstparty = "Reader imports"
 
 [tool.mypy]
 strict = true
 
-    [[tool.mypy.overrides]]
-    module                 = "feedparser"
-    ignore_missing_imports = true
+[[tool.mypy.overrides]]
+module = "feedparser"
+ignore_missing_imports = true
