+++ src/worldenergydata/modules/bsee/data/_by_block/data_from_bin.py
@@ -3,8 +3,6 @@ from copy import deepcopy
 
 import pandas as pd
 
-from worldenergydata.modules.bsee.data.scrapy_block_data import ScrapyRunnerBlock
-
 from assetutilities.common.utilities import is_dir_valid_func
 
 class DataFromBin:
