+++ src/worldenergydata/modules/bsee/data/_by_block/data_from_url.py
@@ -3,7 +3,7 @@ from copy import deepcopy
 
 import pandas as pd
 
-from worldenergydata.modules.bsee.data.scrapy_block_data import ScrapyRunnerBlock
+from worldenergydata.modules.bsee.data._from_url.scrapy_block_data import ScrapyRunnerBlock
 
 from assetutilities.common.utilities import is_dir_valid_func
