+++ src/worldenergydata/modules/bsee/data/bsee_data.py
@@ -1,5 +1,5 @@
-from worldenergydata.modules.bsee.data.well import WellData
-from worldenergydata.modules.bsee.data.production import Production
+from worldenergydata.modules.bsee.data.well.well import WellData
+from worldenergydata.modules.bsee.data.production.production import Production
 from worldenergydata.modules.bsee.data.by_block import Block
 from worldenergydata.modules.bsee.data.refresh.data_refresh import DataRefresh
