+++ src/worldenergydata/modules/bsee/data/by_block.py
@@ -22,8 +22,6 @@ class Block:
         if 'analysis' in cfg and cfg['analysis']['bottom_blocks']:
             cfg = self.add_api12_array_by_block_to_cfg(cfg)
 
-
-
         return cfg
     
     def add_api12_array_by_block_to_cfg(self, cfg):
