+++ src/worldenergydata/modules/bsee/data/production/production.py
@@ -1,5 +1,5 @@
-from worldenergydata.modules.bsee.data.production_data_sources import ProductionDataFromSources
-from worldenergydata.modules.bsee.data.production_data_from_zip import GetProdDataFromZip
+from worldenergydata.modules.bsee.data.production.production_data_sources import ProductionDataFromSources
+from worldenergydata.modules.bsee.data.production.production_data_from_zip import GetProdDataFromZip
 
 production_data_sources = ProductionDataFromSources()
 production_from_zip = GetProdDataFromZip()
