+++ src/worldenergydata/modules/bsee/data/production/production_data_sources.py
@@ -3,8 +3,8 @@ from copy import deepcopy
 
 import pandas as pd
 
-from worldenergydata.modules.bsee.data.scrapy_production_data import ScrapyRunnerProduction
-from worldenergydata.modules.bsee.data.production_data_from_zip import GetProdDataFromZip
+from worldenergydata.modules.bsee.data._from_url.scrapy_production_data import ScrapyRunnerProduction
+from worldenergydata.modules.bsee.data.production.production_data_from_zip import GetProdDataFromZip
 
 from assetutilities.common.utilities import is_dir_valid_func
