+++ src/worldenergydata/modules/bsee/data/refresh/data_refresh.py
@@ -1,7 +1,7 @@
 import logging
 
-from worldenergydata.modules.bsee.data.production_data_from_zip import GetProdDataFromZip
-from worldenergydata.modules.bsee.data.well_from_zip import WellDataFromZip
+from worldenergydata.modules.bsee.data.production.production_data_from_zip import GetProdDataFromZip
+from worldenergydata.modules.bsee.data.well.well_from_zip import WellDataFromZip
 
 from assetutilities.common.utilities import is_dir_valid_func
