+++ src/worldenergydata/modules/bsee/data/well/well_from_zip.py
@@ -5,8 +5,7 @@ from copy import deepcopy
 import logging
 from loguru import logger
 
-from worldenergydata.modules.bsee.data.scrapy_well_data import ScrapyRunnerAPI
-from worldenergydata.modules.bsee.data.by_block.block_data import BlockData
+from worldenergydata.modules.bsee.data._by_block.data_from_url import DataFromURL
 
 from assetutilities.common.utilities import is_dir_valid_func
 from assetutilities.common.yml_utilities import WorkingWithYAML  # noqa
@@ -16,7 +15,7 @@ from assetutilities.modules.zip_utilities.zip_files_to_dataframe import ZipFiles
 wwy = WorkingWithYAML()
 zip_files_to_df = ZipFilestoDf()
 
-block_data = BlockData()
+block_data = DataFromURL()
 
 class WellDataFromZip:
